Pull out mouseAction and scroll
