+++ lib/taiko.js
@@ -1212,47 +1212,9 @@ module.exports.mouseAction = mouseAction;
 
 async function mouseAction(selector, action, coordinates, options = {}) {
   validate();
-  var actions = ['press', 'move', 'release'];
-  if (!actions.includes(selector) && !isSelector(selector) && !isString(selector)) {
-    throw Error('Invalid Selector value passed : ' + selector);
-  }
-  if (!actions.includes(selector)) {
-    const elem = await findFirstElement(selector);
-    if (elem == null) {
-      throw Error('Please provide a valid selector, unable to find element');
-    }
-    await scrollToElement(elem);
-    let rect = await domHandler.getBoundingClientRect(elem.get());
-    coordinates.x = coordinates.x + parseInt(rect.left);
-    coordinates.y = coordinates.y + parseInt(rect.top);
-  } else if (actions.includes(selector)) {
-    coordinates = action;
-    action = selector;
-  }
-  options = setNavigationOptions(options);
-  if (defaultConfig.headful) {
-    await overlayHandler.highlightRect({
-      x: coordinates.x,
-      y: coordinates.y,
-      width: 1,
-      height: 1,
-    });
-  }
-  options = setClickOptions(options, coordinates.x, coordinates.y);
-  await doActionAwaitingNavigation(options, async () => {
-    if (action === 'press') {
-      options.type = 'mousePressed';
-    } else if (action === 'move') {
-      options.type = 'mouseMoved';
-    } else if (action === 'release') {
-      options.type = 'mouseReleased';
-    }
-    await inputHandler.dispatchMouseEvent(options);
-  });
-  descEvent.emit(
-    'success',
-    'Performed mouse ' + action + 'action at {' + coordinates.x + ', ' + coordinates.y + '}',
-  );
+  const { mouseAction } = require('./actions/mouseAction');
+  const desc = mouseAction(selector, action, coordinates, options);
+  descEvent.emit('success', desc);
 }
 
 /**
@@ -1277,44 +1239,9 @@ module.exports.scrollTo = async (selector, options = {}) => {
     await scrollToElement(element);
   });
   if (defaultConfig.headful) {
-    await highlightElement(await findFirstElement(selector));
-  }
-  descEvent.emit('success', 'Scrolled to the ' + description(selector, true));
-};
-
-const scroll = async (e, px, scrollPage, scrollElement, direction) => {
-  e = e || 100;
-  if (Number.isInteger(e)) {
-    const res = await runtimeHandler.runtimeEvaluate(
-      `(${scrollPage}).apply(null, ${JSON.stringify([e])})`,
-    );
-    if (res.result.subtype == 'error') {
-      throw new Error(res.result.description);
-    }
-    return {
-      description: `Scrolled ${direction} the page by ${e} pixels`,
-    };
-  }
-
-  const element = await findFirstElement(e);
-  if (defaultConfig.headful) {
     await highlightElement(element);
   }
-  //TODO: Allow user to set options for scroll
-  const options = setNavigationOptions({});
-  await doActionAwaitingNavigation(options, async () => {
-    const res = await runtimeHandler.runtimeCallFunctionOn(scrollElement, null, {
-      objectId: element.get(),
-      arg: px,
-    });
-    if (res.result.subtype == 'error') {
-      throw new Error(res.result.description);
-    }
-  });
-  descEvent.emit(
-    'success',
-    'Scrolled ' + direction + description(e, true) + 'by ' + px + ' pixels',
-  );
+  descEvent.emit('success', 'Scrolled to the ' + description(selector, true));
 };
 
 /**
@@ -1336,7 +1263,8 @@ const scroll = async (e, px, scrollPage, scrollElement, direction) => {
  */
 module.exports.scrollRight = async (e, px = 100) => {
   validate();
-  return await scroll(
+  const { scroll } = require('./actions/scrollTo');
+  const desc = await scroll(
     e,
     px,
     (px) => window.scrollBy(px, 0),
@@ -1350,6 +1278,7 @@ module.exports.scrollRight = async (e, px = 100) => {
     },
     'right',
   );
+  descEvent.emit('success', desc);
 };
 
 /**
@@ -1371,7 +1300,8 @@ module.exports.scrollRight = async (e, px = 100) => {
  */
 module.exports.scrollLeft = async (e, px = 100) => {
   validate();
-  return await scroll(
+  const { scroll } = require('./actions/scrollTo');
+  const desc = await scroll(
     e,
     px,
     (px) => window.scrollBy(px * -1, 0),
@@ -1385,6 +1315,7 @@ module.exports.scrollLeft = async (e, px = 100) => {
     },
     'left',
   );
+  descEvent.emit('success', desc);
 };
 
 /**
@@ -1406,7 +1337,8 @@ module.exports.scrollLeft = async (e, px = 100) => {
  */
 module.exports.scrollUp = async (e, px = 100) => {
   validate();
-  return await scroll(
+  const { scroll } = require('./actions/scrollTo');
+  const desc = await scroll(
     e,
     px,
     (px) => window.scrollBy(0, px * -1),
@@ -1420,6 +1352,7 @@ module.exports.scrollUp = async (e, px = 100) => {
     },
     'up',
   );
+  descEvent.emit('success', desc);
 };
 
 /**
@@ -1441,7 +1374,8 @@ module.exports.scrollUp = async (e, px = 100) => {
  */
 module.exports.scrollDown = async (e, px = 100) => {
   validate();
-  return await scroll(
+  const { scroll } = require('./actions/scrollTo');
+  const desc = await scroll(
     e,
     px,
     (px) => window.scrollBy(0, px),
@@ -1455,6 +1389,7 @@ module.exports.scrollDown = async (e, px = 100) => {
     },
     'down',
   );
+  descEvent('Success', desc);
 };
 
 /**
